add empty services

+++ DataAccessLayer/AbstractionRepositories/IUserRepository.cs
@@ -9,7 +9,7 @@ namespace DataAccessLayer.Abstraction
     public interface IUserRepository
     {
         public Task<User> GetUserByNameAndPassword(AuthorizationMessage authorizationMessage);
-        public Task<int> CreateUser(User user);
+        public Task<long> CreateUser(User user);
         public Task UpdateUser(User user);
         public Task DeleteUser(User user);
     }
